Fix for generate_changelog to grab from driver page
Previously I had taken the driver version and taken it in to
generate_changelog as an argument in order to append it to what I had
predicted as the release notes page every time.

Inveitably it didn't work so instead I choose here to grab it from the
driver page which links to the release notes, hopefully guaranteeing
that the script does not fail in finding the release notes content.